fix(storybook): fix font not loading in stories
