+++ autogpt_platform/frontend/.storybook/manager-head.html
@@ -1,0 +1,31 @@
+<link rel="preconnect" href="https://fonts.googleapis.com" />
+<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+<link
+  href="https://fonts.googleapis.com/css2?family=Poppins:wght@400;500;600;700&display=swap"
+  rel="stylesheet"
+/>
+
+<style>
+  :root {
+    --font-poppins: "Poppins", sans-serif;
+    --font-geist-sans: system-ui, -apple-system, sans-serif;
+    --font-geist-mono: ui-monospace, monospace;
+  }
+
+  body,
+  #storybook-root {
+    font-family:
+      var(--font-poppins),
+      system-ui,
+      -apple-system,
+      sans-serif;
+  }
+
+  .font-sans {
+    font-family:
+      var(--font-poppins),
+      system-ui,
+      -apple-system,
+      sans-serif;
+  }
+</style>
