--- autogpt_platform/frontend/.storybook/preview.ts
@@ -1,23 +1,0 @@
-import type { Preview } from "@storybook/react";
-import { initialize, mswLoader } from "msw-storybook-addon";
-import "../src/app/globals.css";
-
-// Initialize MSW
-initialize();
-
-const preview: Preview = {
-  parameters: {
-    nextjs: {
-      appDirectory: true,
-    },
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/i,
-      },
-    },
-  },
-  loaders: [mswLoader],
-};
-
-export default preview;
